quick fix for CSS and JS not displaying/ created while loop
Tried re organizing files, this caused the CSS and JS to not display, resorted back to original file locations. Added a while loop to run through the series of prompts needed to create the password with loops for incorrect answers. Also added multiple arrays containing the corresponding elements that will later be included in the password.