new variable and conditional statements
Introduced the new variable "password" to be used as an empty array which will hold the generated password value based off user input. Created if statements with the new variable "password" which will assign the user inputs to the password. Used console.log to verify the correct arrays were being targeted based off the users inputs.