Introduced new variables and changed names, Introduced for loop and a return statement
Removed console.logs. Changed variable "password" to "passwordOutcome" to avoid confusion and to make the code easier to read.  Introduced a for loop that collects the input from the user and randomizes the arrays chosen. With the return statement, the new randomly generated password displays in the text area on the webpage.