Attempt to improve user experience
Upon page load, the code works as intended. However, if the user attempts to generate another random password with different criteria, the previously selected criteria will still remain selected. This was an attempt to clear or reset the user inputs once the function has completed, so that the user could generate as many passwords as they wish, with whatever criteria they desire, without having to refresh the page.